backend/requirements.txt
# PDF Processing
PyMuPDF>=1.24.0
pikepdf>=9.0.0
pdfplumber>=0.11.0
reportlab>=4.2.0
pdf2image>=1.17.0

# OCR
pytesseract>=0.3.10
easyocr>=1.7.0
Pillow>=10.4.0

# NLP and PII Detection
spacy>=3.7.0
transformers>=4.40.0
presidio-analyzer>=2.2.300
presidio-anonymizer>=2.2.300
phonenumbers>=8.13.0

# Computer Vision
opencv-python>=4.10.0
pyzbar>=0.1.9
scikit-image>=0.24.0
numpy>=1.26.0,<2.0

# Utilities
pydantic>=2.8.0
loguru>=0.7.0
python-dotenv>=1.0.0

# Development
pytest>=8.0.0
pytest-cov>=5.0.0
black>=24.0.0
flake8>=7.0.0
mypy>=1.10.0

# Packaging
pyinstaller>=6.10.0
